Update package names
