+++ src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/LocatorBundle.kt
@@ -5,9 +5,9 @@ import org.jetbrains.annotations.NonNls
 import org.jetbrains.annotations.PropertyKey
 
 @NonNls
-private const val BUNDLE = "messages.MyBundle"
+private const val BUNDLE = "messages.LocatorBundle"
 
-object MyBundle : DynamicBundle(BUNDLE) {
+object LocatorBundle : DynamicBundle(BUNDLE) {
 
     @JvmStatic
     fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =
