+++ src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/git/cli/EmptyCliResponse.kt
@@ -1,0 +1,4 @@
+package com.github.y0ung3r.gitglobalhookslocator.git.cli
+
+class EmptyCliResponse
+    : CliResponse("")
