+++ src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/git/cli/NotFoundCliResponse.kt
@@ -1,4 +1,4 @@
-package com.github.y0ung3r.gitglobalhookslocator.services.git.cli
+package com.github.y0ung3r.gitglobalhookslocator.git.cli
 
 class NotFoundCliResponse(details: String)
     : CliResponse(details)
