+++ src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/git/cli/interfaces/CliCommandExecutor.kt
@@ -1,0 +1,7 @@
+package com.github.y0ung3r.gitglobalhookslocator.git.cli.interfaces
+
+import com.github.y0ung3r.gitglobalhookslocator.git.cli.CliResponse
+
+interface CliCommandExecutor {
+    fun execute(processBuilder: ProcessBuilder): CliResponse
+}
