+++ src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/git/exceptions/GitCommandNotFoundException.kt
@@ -1,4 +1,4 @@
-package com.github.y0ung3r.gitglobalhookslocator.services.git.exceptions
+package com.github.y0ung3r.gitglobalhookslocator.git.exceptions
 
 class GitCommandNotFoundException(vararg params: String)
     : Exception("Native Git command \"git ${params.joinToString(" ")}\" not found")
