+++ src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/git/exceptions/GitIsNotInstalledException.kt
@@ -1,4 +1,4 @@
-package com.github.y0ung3r.gitglobalhookslocator.services.git.exceptions
+package com.github.y0ung3r.gitglobalhookslocator.git.exceptions
 
 class GitIsNotInstalledException
     : Exception("Git is not installed")
