+++ src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/git/exceptions/GitVersionIsNotSupportedException.kt
@@ -1,0 +1,6 @@
+package com.github.y0ung3r.gitglobalhookslocator.git.exceptions
+
+import com.github.y0ung3r.gitglobalhookslocator.git.Git
+
+class GitVersionIsNotSupportedException
+    : Exception("The plugin requires Git version above ${Git.minRequiredVersion}")
