+++ src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/git/exceptions/HookNotFoundException.kt
@@ -1,4 +1,4 @@
-package com.github.y0ung3r.gitglobalhookslocator.services.git.exceptions
+package com.github.y0ung3r.gitglobalhookslocator.git.exceptions
 
 class HookNotFoundException(fileName: String)
     : Exception("Global hook \"$fileName\" not found")
