+++ src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/git/exceptions/ProvidedSemanticVersionIsInvalidException.kt
@@ -1,4 +1,4 @@
-package com.github.y0ung3r.gitglobalhookslocator.services.git.exceptions
+package com.github.y0ung3r.gitglobalhookslocator.git.exceptions
 
 class ProvidedSemanticVersionIsInvalidException(version: String)
     : Exception("The provided version does not conform to semantic versioning: $version")
