--- src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/services/git/GitResponse.kt
@@ -1,9 +1,0 @@
-package com.github.y0ung3r.gitglobalhookslocator.services.git
-
-import com.github.y0ung3r.gitglobalhookslocator.services.git.cli.CliResponse
-import com.github.y0ung3r.gitglobalhookslocator.services.git.cli.EmptyCliResponse
-
-class GitResponse(private val cliResponse: CliResponse) {
-    val value = cliResponse.value
-    fun isEmpty() = cliResponse is EmptyCliResponse
-}
