--- src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/services/git/cli/EmptyCliResponse.kt
@@ -1,4 +1,0 @@
-package com.github.y0ung3r.gitglobalhookslocator.services.git.cli
-
-class EmptyCliResponse
-    : CliResponse("")
