--- src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/services/git/cli/interfaces/CliCommandExecutor.kt
@@ -1,7 +1,0 @@
-package com.github.y0ung3r.gitglobalhookslocator.services.git.cli.interfaces
-
-import com.github.y0ung3r.gitglobalhookslocator.services.git.cli.CliResponse
-
-interface CliCommandExecutor {
-    fun execute(processBuilder: ProcessBuilder): CliResponse
-}
