--- src/main/kotlin/com/github/y0ung3r/gitglobalhookslocator/toolWindow/MyToolWindowFactory.kt
@@ -1,81 +1,0 @@
-package com.github.y0ung3r.gitglobalhookslocator.toolWindow
-
-import com.intellij.openapi.components.service
-import com.intellij.openapi.diagnostic.thisLogger
-import com.intellij.openapi.project.Project
-import com.intellij.openapi.wm.ToolWindow
-import com.intellij.openapi.wm.ToolWindowFactory
-import com.intellij.ui.components.JBLabel
-import com.intellij.ui.components.JBPanel
-import com.intellij.ui.content.ContentFactory
-import com.github.y0ung3r.gitglobalhookslocator.MyBundle
-import com.github.y0ung3r.gitglobalhookslocator.services.MyProjectService
-import com.intellij.openapi.vcs.CheckinProjectPanel
-import com.intellij.openapi.vcs.changes.CommitContext
-import com.intellij.openapi.vcs.checkin.CheckinHandler
-import com.intellij.openapi.vcs.checkin.CheckinHandlerFactory
-import com.intellij.openapi.vcs.ui.RefreshableOnComponent
-import com.intellij.ui.NonFocusableCheckBox
-import java.awt.GridLayout
-import javax.swing.JButton
-import javax.swing.JComponent
-import javax.swing.JPanel
-
-class GlobalHooksLocatorFactory : CheckinHandlerFactory() {
-    override fun createHandler(panel: CheckinProjectPanel, commitContext: CommitContext): CheckinHandler
-            = GlobalHooksLocator()
-}
-
-class GlobalHooksLocator : CheckinHandler() {
-    override fun getBeforeCheckinConfigurationPanel(): RefreshableOnComponent? {
-        val commonCheckBox = NonFocusableCheckBox("Disable all Global hooks")
-
-        return object : RefreshableOnComponent {
-            override fun saveState() {
-
-            }
-
-            override fun restoreState() {
-
-            }
-
-            override fun getComponent(): JComponent {
-                val panel = JPanel(GridLayout(1, 0))
-                panel.add(commonCheckBox)
-                return panel
-            }
-
-        }
-    }
-}
-
-class MyToolWindowFactory : ToolWindowFactory {
-
-    init {
-        thisLogger().warn("Don't forget to remove all non-needed sample code files with their corresponding registration entries in `plugin.xml`.")
-    }
-
-    override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
-        val myToolWindow = MyToolWindow(toolWindow)
-        val content = ContentFactory.getInstance().createContent(myToolWindow.getContent(), null, false)
-        toolWindow.contentManager.addContent(content)
-    }
-
-    override fun shouldBeAvailable(project: Project) = true
-
-    class MyToolWindow(toolWindow: ToolWindow) {
-
-        private val service = toolWindow.project.service<MyProjectService>()
-
-        fun getContent() = JBPanel<JBPanel<*>>().apply {
-            val label = JBLabel(MyBundle.message("randomLabel", "?"))
-
-            add(label)
-            add(JButton(MyBundle.message("shuffle")).apply {
-                addActionListener {
-                    label.text = MyBundle.message("randomLabel", service.getRandomNumber())
-                }
-            })
-        }
-    }
-}
