+++ src/main/resources/META-INF/plugin.xml
@@ -1,19 +1,14 @@
 <!-- Plugin Configuration File. Read more: https://plugins.jetbrains.com/docs/intellij/plugin-configuration-file.html -->
 <idea-plugin>
     <id>com.github.y0ung3r.gitglobalhookslocator</id>
-    <name>GitGlobalHooksLocator Template</name>
+    <name>Git Global Hooks Locator</name>
     <vendor>y0ung3r</vendor>
 
     <depends>com.intellij.modules.platform</depends>
 
-    <resource-bundle>messages.MyBundle</resource-bundle>
+    <resource-bundle>messages.LocatorBundle</resource-bundle>
 
     <extensions defaultExtensionNs="com.intellij">
-        <toolWindow factoryClass="com.github.y0ung3r.gitglobalhookslocator.toolWindow.MyToolWindowFactory" id="MyToolWindow"/>
-        <checkinHandlerFactory implementation="com.github.y0ung3r.gitglobalhookslocator.toolWindow.GlobalHooksLocatorFactory" />
+        <toolWindow factoryClass="com.github.y0ung3r.gitglobalhookslocator.ui.toolWindow.GitGlobalHooksLocatorWindowFactory" id="GitGlobalHooksLocatorToolWindow"/>
     </extensions>
-
-    <applicationListeners>
-        <listener class="com.github.y0ung3r.gitglobalhookslocator.listeners.MyApplicationActivationListener" topic="com.intellij.openapi.application.ApplicationActivationListener"/>
-    </applicationListeners>
 </idea-plugin>
