+++ src/main/resources/messages/LocatorBundle.properties
@@ -1,0 +1,1 @@
+locator.ui.toolWindow.disableAll=Disable all Git global hooks
