--- src/main/resources/messages/MyBundle.properties
@@ -1,3 +1,0 @@
-projectService=Project service: {0}
-randomLabel=The random number is: {0}
-shuffle=Shuffle
