+++ src/test/kotlin/com/github/y0ung3r/gitglobalhookslocator/gitTests/GetGlobalHooksPathTests.kt
@@ -1,9 +1,9 @@
 package com.github.y0ung3r.gitglobalhookslocator.gitTests
 
 import com.github.y0ung3r.gitglobalhookslocator.gitTests.testEngine.RespondInterchangeably
-import com.github.y0ung3r.gitglobalhookslocator.services.git.Git
-import com.github.y0ung3r.gitglobalhookslocator.services.git.cli.CliResponse
-import com.github.y0ung3r.gitglobalhookslocator.services.git.extensions.getGlobalHooksPath
+import com.github.y0ung3r.gitglobalhookslocator.git.Git
+import com.github.y0ung3r.gitglobalhookslocator.git.cli.CliResponse
+import com.github.y0ung3r.gitglobalhookslocator.git.extensions.getGlobalHooksPath
 import org.junit.Assert.assertEquals
 import org.junit.Test
