+++ src/test/kotlin/com/github/y0ung3r/gitglobalhookslocator/gitTests/MinGitVersionTests.kt
@@ -1,9 +1,9 @@
 package com.github.y0ung3r.gitglobalhookslocator.gitTests
 
 import com.github.y0ung3r.gitglobalhookslocator.gitTests.testEngine.RespondOnce
-import com.github.y0ung3r.gitglobalhookslocator.services.git.Git
-import com.github.y0ung3r.gitglobalhookslocator.services.git.cli.CliResponse
-import com.github.y0ung3r.gitglobalhookslocator.services.git.exceptions.GitVersionIsNotSupportedException
+import com.github.y0ung3r.gitglobalhookslocator.git.Git
+import com.github.y0ung3r.gitglobalhookslocator.git.cli.CliResponse
+import com.github.y0ung3r.gitglobalhookslocator.git.exceptions.GitVersionIsNotSupportedException
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
