+++ src/test/kotlin/com/github/y0ung3r/gitglobalhookslocator/gitTests/testEngine/RespondInterchangeably.kt
@@ -1,7 +1,7 @@
 package com.github.y0ung3r.gitglobalhookslocator.gitTests.testEngine
 
-import com.github.y0ung3r.gitglobalhookslocator.services.git.cli.CliResponse
-import com.github.y0ung3r.gitglobalhookslocator.services.git.cli.interfaces.CliCommandExecutor
+import com.github.y0ung3r.gitglobalhookslocator.git.cli.CliResponse
+import com.github.y0ung3r.gitglobalhookslocator.git.cli.interfaces.CliCommandExecutor
 
 class RespondInterchangeably(vararg responses: CliResponse) : CliCommandExecutor {
     private val iterator = responses.iterator()
